+++ ReleaseNotes.md
@@ -5,6 +5,7 @@ Latest update: December 16th 2017
 ## Bug Fixes
 
 * Change the dialog.
+* Append final to JButton in run().
 
 # Release v0.9.1.0